Extract proptest strategies out into test module
