+++ common/http/Cargo.toml
@@ -32,9 +32,9 @@ vec-collections = "0.3.5"
 smallvec = "1.4"
 
 [dev-dependencies]
-async-trait = "0.1"
 tokio = { version = "1", features = ["rt", "macros"] }
 tokio-test = "0.4"
 rand = "0.8"
 num_cpus = "1.0"
-proptest = "0.10"
+proptest = "1"
+test_types = { package = "types", path = "../test/types" }
