+++ common/http/src/batch.rs
@@ -210,205 +210,17 @@ mod tests {
 
     use super::*;
 
+    use std::collections::HashMap;
     use std::io::Read;
     use std::time::{Duration, Instant};
 
-    use async_trait::async_trait;
+    use futures::stream;
+    use futures::FutureExt;
 
-    use futures::{stream, FutureExt, StreamExt};
-
-    use proptest::collection::hash_map;
-    use proptest::option::of;
     use proptest::prelude::*;
-    use proptest::string::string_regex;
 
-    use crate::types::body::{KeyValueMap, Line, LineMeta};
-    use crate::types::serialize::{
-        SerializeI64, SerializeMap, SerializeStr, SerializeUtf8, SerializeValue,
-    };
-
-    #[derive(Clone, Debug, PartialEq)]
-    struct OffsetLine {
-        line: Line,
-    }
-
-    impl OffsetLine {
-        fn new(line: Line) -> Self {
-            OffsetLine { line }
-        }
-    }
-
-    impl GetOffset for &OffsetLine {
-        fn get_offset(&self) -> Option<u64> {
-            None
-        }
-        fn get_key(&self) -> Option<u64> {
-            None
-        }
-    }
-
-    #[async_trait]
-    impl IngestLineSerialize<String, bytes::Bytes, std::collections::HashMap<String, String>>
-        for &OffsetLine
-    {
-        type Ok = ();
-
-        fn has_annotations(&self) -> bool {
-            self.line.get_annotations().is_some()
-        }
-        async fn annotations<'b, S>(
-            &mut self,
-            writer: &mut S,
-        ) -> Result<Self::Ok, IngestLineSerializeError>
-        where
-            S: SerializeMap<'b, HashMap<String, String>> + std::marker::Send,
-        {
-            (&self.line).annotations(writer).await
-        }
-        fn has_app(&self) -> bool {
-            self.line.get_app().is_some()
-        }
-        async fn app<S>(&mut self, writer: &mut S) -> Result<(), IngestLineSerializeError>
-        where
-            S: SerializeStr<String> + std::marker::Send,
-        {
-            (&self.line).app(writer).await
-        }
-        fn has_env(&self) -> bool {
-            self.line.get_env().is_some()
-        }
-        async fn env<S>(&mut self, writer: &mut S) -> Result<(), IngestLineSerializeError>
-        where
-            S: SerializeStr<String> + std::marker::Send,
-        {
-            (&self.line).env(writer).await
-        }
-        fn has_file(&self) -> bool {
-            self.line.get_file().is_some()
-        }
-        async fn file<S>(&mut self, writer: &mut S) -> Result<(), IngestLineSerializeError>
-        where
-            S: SerializeStr<String> + std::marker::Send,
-        {
-            (&self.line).file(writer).await
-        }
-        fn has_host(&self) -> bool {
-            self.line.get_host().is_some()
-        }
-        async fn host<S>(&mut self, writer: &mut S) -> Result<(), IngestLineSerializeError>
-        where
-            S: SerializeStr<String> + std::marker::Send,
-        {
-            (&self.line).host(writer).await
-        }
-        fn has_labels(&self) -> bool {
-            self.line.get_labels().is_some()
-        }
-        async fn labels<'b, S>(&mut self, writer: &mut S) -> Result<(), IngestLineSerializeError>
-        where
-            S: SerializeMap<'b, HashMap<String, String>> + std::marker::Send,
-        {
-            (&self.line).labels(writer).await
-        }
-        fn has_level(&self) -> bool {
-            self.line.get_level().is_some()
-        }
-        async fn level<S>(&mut self, writer: &mut S) -> Result<(), IngestLineSerializeError>
-        where
-            S: SerializeStr<String> + std::marker::Send,
-        {
-            (&self.line).level(writer).await
-        }
-        fn has_meta(&self) -> bool {
-            self.line.get_meta().is_some()
-        }
-        async fn meta<S>(&mut self, writer: &mut S) -> Result<(), IngestLineSerializeError>
-        where
-            S: SerializeValue + std::marker::Send,
-        {
-            (&self.line).meta(writer).await
-        }
-        async fn line<S>(&mut self, writer: &mut S) -> Result<(), IngestLineSerializeError>
-        where
-            S: SerializeUtf8<bytes::Bytes> + std::marker::Send,
-        {
-            (&self.line).line(writer).await
-        }
-        async fn timestamp<S>(&mut self, writer: &mut S) -> Result<(), IngestLineSerializeError>
-        where
-            S: SerializeI64 + std::marker::Send,
-        {
-            (&self.line).timestamp(writer).await
-        }
-        fn field_count(&self) -> usize {
-            (&self.line).field_count()
-        }
-    }
-
-    fn key_value_map_st(max_entries: usize) -> impl Strategy<Value = KeyValueMap> {
-        hash_map(
-            string_regex(".{1,64}").unwrap(),
-            string_regex(".{1,64}").unwrap(),
-            0..max_entries,
-        )
-        .prop_map(move |c| {
-            let mut kv_map = KeyValueMap::new();
-            for (k, v) in c.into_iter() {
-                kv_map = kv_map.add(k, v);
-            }
-            kv_map
-        })
-    }
-
-    //recursive JSON type
-    fn json_st(depth: u32) -> impl Strategy<Value = serde_json::Value> {
-        let leaf = prop_oneof![
-            Just(serde_json::Value::Null),
-            any::<bool>().prop_map(|o| serde_json::to_value(o).unwrap()),
-            any::<f64>().prop_map(|o| serde_json::to_value(o).unwrap()),
-            ".{1,64}".prop_map(|o| serde_json::to_value(o).unwrap()),
-        ];
-        leaf.prop_recursive(depth, 256, 10, |inner| {
-            prop_oneof![
-                prop::collection::vec(inner.clone(), 0..10)
-                    .prop_map(|o| serde_json::to_value(o).unwrap()),
-                prop::collection::hash_map(".*", inner, 0..10)
-                    .prop_map(|o| serde_json::to_value(o).unwrap()),
-            ]
-        })
-    }
-    fn line_st() -> impl Strategy<Value = OffsetLine> {
-        (
-            of(key_value_map_st(5)),
-            of(string_regex(".{1,64}").unwrap()),
-            of(string_regex(".{1,64}").unwrap()),
-            of(string_regex(".{1,64}").unwrap()),
-            of(string_regex(".{1,64}").unwrap()),
-            of(key_value_map_st(5)),
-            of(string_regex(".{1,64}").unwrap()),
-            of(json_st(3)),
-            string_regex(".{1,64}").unwrap(),
-            (0..i64::MAX),
-        )
-            .prop_map(
-                |(annotations, app, env, file, host, labels, level, meta, line, timestamp)| {
-                    OffsetLine {
-                        line: Line {
-                            annotations,
-                            app,
-                            env,
-                            file,
-                            host,
-                            labels,
-                            level,
-                            meta,
-                            line,
-                            timestamp,
-                        },
-                    }
-                },
-            )
-    }
+    use crate::types::body::Line;
+    use test_types::strategies::{line_st, OffsetLine};
 
     #[tokio::test]
     async fn messages_pass_through() {
