+++ common/test/types/Cargo.toml
@@ -7,4 +7,10 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+async-trait = "0.1"
+bytes = "1"
+futures = "0.3"
+logdna-client = { git = "https://github.com/logdna/logdna-rust.git", branch="len", version = "0.5" }
 proptest = "1"
+serde_json = "1"
+state = { package = "state", path = "../../state" }
